+++ src/core/api/index.ts
@@ -486,7 +486,7 @@ class Player extends EventEmitter<PLAYER_EVENT_STRINGS, any> {
     // See: https://bugzilla.mozilla.org/show_bug.cgi?id=1194624
     videoElement.preload = "auto";
 
-    this.version = /*PLAYER_VERSION*/"3.4.0";
+    this.version = /*PLAYER_VERSION*/"3.4.1";
     this.log = log;
     this.state = "STOPPED";
     this.videoElement = videoElement;
@@ -1510,13 +1510,13 @@ class Player extends EventEmitter<PLAYER_EVENT_STRINGS, any> {
    * Returns every available audio tracks for the current Period.
    * @returns {Array.<Object>|null}
    */
-  getAvailableAudioTracks() : ILMAudioTrackListItem[] | null {
+  getAvailableAudioTracks() : ILMAudioTrackListItem[] {
     if (!this._priv_contentInfos) {
-      return null;
+      return [];
     }
     const { currentPeriod } = this._priv_contentInfos;
     if (!this._priv_languageManager || !currentPeriod) {
-      return null;
+      return [];
     }
     return this._priv_languageManager.getAvailableAudioTracks(currentPeriod);
   }
@@ -1525,13 +1525,13 @@ class Player extends EventEmitter<PLAYER_EVENT_STRINGS, any> {
    * Returns every available text tracks for the current Period.
    * @returns {Array.<Object>|null}
    */
-  getAvailableTextTracks() : ILMTextTrackListItem[] | null {
+  getAvailableTextTracks() : ILMTextTrackListItem[] {
     if (!this._priv_contentInfos) {
-      return null;
+      return [];
     }
     const { currentPeriod } = this._priv_contentInfos;
     if (!this._priv_languageManager || !currentPeriod) {
-      return null;
+      return [];
     }
     return this._priv_languageManager.getAvailableTextTracks(currentPeriod);
   }