+++ src/net/dash/index.ts
@@ -80,14 +80,15 @@ export default function(
     },
 
     parser(
-      { response, url } : IManifestParserArguments<Document|string>
+      { response, url: reqURL } : IManifestParserArguments<Document|string>
     ) : IManifestParserObservable {
+      const url = response.url == null ? reqURL : response.url;
       const data = typeof response.responseData === "string" ?
         new DOMParser().parseFromString(response.responseData, "text/xml") :
         response.responseData;
       return Observable.of({
-        manifest: dashManifestParser(data, url/*, contentProtectionParser*/),
-        url: response.url,
+        manifest: dashManifestParser(data, url),
+        url,
       });
     },
   };