+++ tests/integration/scenarios/dash_live.js
@@ -3,7 +3,7 @@ import sinon from "sinon";
 import RxPlayer from "../../../src";
 import { mockManifestRequest } from "../utils/mock_requests.js";
 import sleep from "../utils/sleep.js";
-import Mock from "../mocks/dash_dynamic_SegmentTimeline.js";
+import Mock from "../mocks/dash-if_segment-timeline.js";
 
 describe("dash live SegmentTimeline content", function () {
   let player;
@@ -34,10 +34,10 @@ describe("dash live SegmentTimeline content", function () {
     expect(manifest.transport).to.equal("dash");
     expect(typeof manifest.id).to.equal("string");
     expect(manifest.isLive).to.equal(true);
-    expect(manifest.suggestedPresentationDelay).to.equal(15);
-    expect(manifest.timeShiftBufferDepth).to.equal(50);
+    expect(manifest.suggestedPresentationDelay).to.equal(10);
+    expect(manifest.timeShiftBufferDepth).to.equal(300);
     expect(manifest.getUrl()).to.equal(Mock.manifest.url);
-    expect(manifest.availabilityStartTime).to.equal(1493225291);
+    expect(manifest.availabilityStartTime).to.equal(0);
 
     const adaptations = manifest.adaptations;
 
@@ -49,95 +49,127 @@ describe("dash live SegmentTimeline content", function () {
     expect(typeof adaptations.audio[0].id).to.equal("string");
     expect(adaptations.audio[0].id).to.not.equal(adaptations.video[0].id);
     expect(adaptations.audio[0].representations.length).to.equal(1);
-    expect(adaptations.audio[0].getAvailableBitrates()).to.eql([96257]);
+    expect(adaptations.audio[0].getAvailableBitrates()).to.eql([48000]);
 
     expect(adaptations.video.length).to.equal(1);
     expect(adaptations.video[0].type).to.equal("video");
-    expect(adaptations.video[0].getAvailableBitrates()).to.eql([601392]);
+    expect(adaptations.video[0].getAvailableBitrates()).to.eql([300000]);
 
     const audioRepresentation = adaptations.audio[0].representations[0];
-    expect(audioRepresentation.baseURL).to.equal("https://wowzaec2demo.streamlock.net/live/_definst_/bigbuckbunny/");
-    expect(audioRepresentation.bitrate).to.equal(96257);
+    expect(audioRepresentation.baseURL).to.equal("http://vm2.dashif.org/livesim-dev/segtimeline_1/testpic_6s/");
+    expect(audioRepresentation.bitrate).to.equal(48000);
     expect(audioRepresentation.codec).to.equal("mp4a.40.2");
-    expect(audioRepresentation.id).to.equal("p0aa0br96257");
+    expect(audioRepresentation.id).to.equal("A48");
     expect(audioRepresentation.mimeType).to.equal("audio/mp4");
     expect(typeof audioRepresentation.index).to.equal("object");
 
     const audioRepresentationIndex = audioRepresentation.index;
     const initAudioSegment = audioRepresentationIndex.getInitSegment();
-    expect(initAudioSegment.media).to.equal("chunk_ctaudio_cfm4s_rid$RepresentationID$_cinit_w925796611_mpd.m4s");
     expect(initAudioSegment.id).to.equal("init");
 
-    const nextAudioSegment1 = audioRepresentationIndex.getSegments(3023671, 10);
+    const nextAudioSegment1 = audioRepresentationIndex
+      .getSegments(1527507762, 5);
     expect(nextAudioSegment1.length).to.equal(1);
-    expect(nextAudioSegment1[0].duration).to.equal(479232);
-    expect(nextAudioSegment1[0].id).to.equal("145136233968");
+    expect(nextAudioSegment1[0].duration).to.equal(288768);
+    expect(nextAudioSegment1[0].id).to.equal("73320372578304");
     expect(nextAudioSegment1[0].isInit).to.equal(false);
-    expect(nextAudioSegment1[0].media).to.equal("chunk_ctaudio_cfm4s_rid$RepresentationID$_cs$Time$_w925796611_mpd.m4s");
-    expect(nextAudioSegment1[0].time).to.equal(145136233968);
+    expect(nextAudioSegment1[0].time).to.equal(73320372578304);
     expect(nextAudioSegment1[0].timescale).to.equal(48000);
 
-    const nextAudioSegment2 = audioRepresentationIndex.getSegments(3023671, 11);
+    const nextAudioSegment2 = audioRepresentationIndex
+      .getSegments(1527507762, 11);
     expect(nextAudioSegment2.length).to.equal(2);
-    expect(nextAudioSegment2[1].duration).to.equal(480288);
-    expect(nextAudioSegment2[1].id).to.equal("145136713200");
+    expect(nextAudioSegment2[1].duration).to.equal(287744);
+    expect(nextAudioSegment2[1].id).to.equal("73320372867072");
     expect(nextAudioSegment2[1].isInit).to.equal(false);
-    expect(nextAudioSegment2[1].media).to.equal("chunk_ctaudio_cfm4s_rid$RepresentationID$_cs$Time$_w925796611_mpd.m4s");
-    expect(nextAudioSegment2[1].time).to.equal(145136713200);
+    expect(nextAudioSegment2[1].time).to.equal(73320372867072);
     expect(nextAudioSegment2[1].timescale).to.equal(48000);
 
-    expect(audioRepresentationIndex.getSegments(3023671, 41).length)
-      .to.equal(5);
-    expect(audioRepresentationIndex.getSegments(3023671, 300000000000).length)
-      .to.equal(5);
+    expect(audioRepresentationIndex.getSegments(1527507762, 294).length)
+      .to.equal(49);
+    expect(audioRepresentationIndex.getSegments(1527507762, 295).length)
+      .to.equal(50);
+    expect(
+      audioRepresentationIndex.getSegments(1527507762, 300000000000).length
+    ).to.equal(50);
 
     const videoRepresentation = adaptations.video[0].representations[0];
-    expect(videoRepresentation.baseURL).to.equal("https://wowzaec2demo.streamlock.net/live/_definst_/bigbuckbunny/");
-    expect(videoRepresentation.bitrate).to.equal(601392);
-    expect(videoRepresentation.codec).to.equal("avc1.42c015");
-    expect(videoRepresentation.id).to.equal("p0va0br601392");
-    expect(videoRepresentation.height).to.equal(288);
-    expect(videoRepresentation.width).to.equal(512);
+    expect(videoRepresentation.baseURL).to.equal("http://vm2.dashif.org/livesim-dev/segtimeline_1/testpic_6s/");
+    expect(videoRepresentation.bitrate).to.equal(300000);
+    expect(videoRepresentation.codec).to.equal("avc1.64001e");
+    expect(videoRepresentation.id).to.equal("V300");
+    expect(videoRepresentation.height).to.equal(360);
+    expect(videoRepresentation.width).to.equal(640);
     expect(videoRepresentation.mimeType).to.equal("video/mp4");
     expect(typeof videoRepresentation.index).to.equal("object");
 
     const videoRepresentationIndex = videoRepresentation.index;
-    const initVideoSegment = videoRepresentationIndex.getInitSegment();
-    expect(initVideoSegment.media).to.equal("chunk_ctvideo_cfm4s_rid$RepresentationID$_cinit_w925796611_mpd.m4s");
 
-    const nextVideoSegment1 = videoRepresentationIndex.getSegments(3023671, 10);
+    const nextVideoSegment1 = videoRepresentationIndex
+      .getSegments(1527507762, 5);
     expect(nextVideoSegment1.length).to.equal(1);
-    expect(nextVideoSegment1[0].duration).to.equal(900000);
-    expect(nextVideoSegment1[0].id).to.equal("272130436800");
+    expect(nextVideoSegment1[0].duration).to.equal(540000);
+    expect(nextVideoSegment1[0].id).to.equal("137475698580000");
     expect(nextVideoSegment1[0].isInit).to.equal(false);
-    expect(nextVideoSegment1[0].media).to.equal("chunk_ctvideo_cfm4s_rid$RepresentationID$_cs$Time$_w925796611_mpd.m4s");
-    expect(nextVideoSegment1[0].time).to.equal(272130436800);
+    expect(nextVideoSegment1[0].time).to.equal(137475698580000);
     expect(nextVideoSegment1[0].timescale).to.equal(90000);
 
-    const nextVideoSegment2 = videoRepresentationIndex.getSegments(3023671, 11);
+    const nextVideoSegment2 = videoRepresentationIndex
+      .getSegments(1527507762, 11);
     expect(nextVideoSegment2.length).to.equal(2);
-    expect(nextVideoSegment2[1].duration).to.equal(900000);
-    expect(nextVideoSegment2[1].id).to.equal("272131336800");
+    expect(nextVideoSegment2[1].duration).to.equal(540000);
+    expect(nextVideoSegment2[1].id).to.equal("137475699120000");
     expect(nextVideoSegment2[1].isInit).to.equal(false);
-    expect(nextVideoSegment2[1].media).to.equal("chunk_ctvideo_cfm4s_rid$RepresentationID$_cs$Time$_w925796611_mpd.m4s");
-    expect(nextVideoSegment2[1].time).to.equal(272131336800);
+    expect(nextVideoSegment2[1].time).to.equal(137475699120000);
     expect(nextVideoSegment2[1].timescale).to.equal(90000);
 
-    expect(videoRepresentationIndex.getSegments(3023671, 41).length)
-      .to.equal(5);
-    expect(videoRepresentationIndex.getSegments(3023671, 300000000000).length)
-      .to.equal(5);
-
+    expect(videoRepresentationIndex.getSegments(1527507762, 294).length)
+      .to.equal(49);
+    expect(videoRepresentationIndex.getSegments(1527507762, 295).length)
+      .to.equal(50);
+    expect(
+      videoRepresentationIndex.getSegments(1527507762, 300000000000).length
+    ).to.equal(50);
 
     expect(fakeServer.requests.length).to.equal(3);
-
     const requestsDone = [
       fakeServer.requests[1].url,
       fakeServer.requests[2].url,
     ];
-
     expect(requestsDone).to.include(Mock.audio[0].init.url);
     expect(requestsDone).to.include(Mock.video[0].init.url);
+
+    // TODO Do the init segment and the first needed segment in parallel
+    // expect(fakeServer.requests.length).to.equal(5);
+
+    // const requestsDone = [
+    //   fakeServer.requests[1].url,
+    //   fakeServer.requests[2].url,
+    //   fakeServer.requests[3].url,
+    //   fakeServer.requests[4].url,
+    // ];
+
+    // expect(requestsDone).to.include(Mock.audio[0].init.url);
+    // expect(requestsDone).to.include(Mock.video[0].init.url);
+
+    // let audioSegmentsDownloaded = 0;
+    // let videoSegmentsDownloaded = 0;
+
+    // const audioSegmentsUrl = Mock.audio[0].segments.map(s => s.url);
+    // const videoSegmentsUrl = Mock.video[0].segments.map(s => s.url);
+    // for (let i = 0; i < requestsDone.length; i++) {
+    //   const url = requestsDone[i];
+    //   if (audioSegmentsUrl.includes(url)) {
+    //     audioSegmentsDownloaded++;
+    //   }
+
+    //   if (videoSegmentsUrl.includes(url)) {
+    //     videoSegmentsDownloaded++;
+    //   }
+    // }
+
+    // expect(audioSegmentsDownloaded).to.equal(1);
+    // expect(videoSegmentsDownloaded).to.equal(1);
   });
 
   it("should list the right bitrates", async function () {
@@ -148,8 +180,8 @@ describe("dash live SegmentTimeline content", function () {
     fakeServer.respond();
     await sleep(5);
 
-    expect(player.getAvailableAudioBitrates()).to.eql([96257]);
-    expect(player.getAvailableVideoBitrates()).to.eql([601392]);
+    expect(player.getAvailableAudioBitrates()).to.eql([48000]);
+    expect(player.getAvailableVideoBitrates()).to.eql([300000]);
   });
 
   it("should list the right languages", async function () {