+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v3.4.1 (2018/05/29)
+
+### Bug fixes
+
+  - buffer: fix several bugs happening when calling ``endOfStream`` to announce the end of the current content. Especially prevalent on Chrome.
+  - net: use redirected URL as a base for further requests when the manifest request led to a HTTP redirect.
+  - vtt/srt: ignore silently (do not throw) when an unknown block has been detected in a vtt or srt file
+  - vtt/srt: support styling spans (like b, i and u XML tags) spanning multiple lines
+  - api: ``getAvailableTextTracks`` and ``getAvailableAudioTracks`` now always return an array (and never null) as anounced in the API documentation
+  - api: set default log level to ``"NONE"`` instead of ``"INFO"``
+  - misc: remove development-only code from the non-minified code
+
+### Other improvements
+
+  - misc: move some dev dependencies from ``dependencies`` to ``devDependencies`` in ``package.json``
+
+
 ## v3.4.0 (2018/05/17)
 
 ### Added